IPDL processing and cross-referencing enhancments
This includes a change to .gitmodules so that my very minor mod to the IPDL
parser to cause it to skip files that fail to parse instead of fatally
breaking can be used.  I was hoping to push this to the mozsearch IPDL parser
repo as a separate branch to avoid this, but I lack permissions on the repo.