change instance group mechanism, make methods and tables work
Instance groups were initially baked into the node types for classes, but
this felt weird/awkward to use and somewhat asymmetric.  Now various levels
of hierarchy can be marked with specific instance groups which can be easily
duplicated and thereby allow a diagram to be annotated after the fact.
(Versus having to re-create block types.)

Methods are rendered now and the overhauls to the node determination work, but
the presentation of the table record labels needs some work.  Specifically:
- The class name at the top wants to be emphasized somewhat.  (Bold?)
- data-symbols is only being annotated onto the full table
- The method symbols should have their classname stripped off.  (Also, the
  methods should potentially be qualified in terms of their parent from the
  inherent nesting?)