Checkpoint UI with vertical menu on the right.
See concurrent screenshot.  It's not working.

The splitter layout where the top box also contains (and thereby limits
vertical space of) the tabbed UI isn't working.  Having the splitter
location be per-tab would be an improvement, but it's still the case
that the "chrome" needs to exist outside whatever's shown in the top.

The most viable approach seems like it would be a vscode style vertical
iconic toolbar down the left side of the screen with hover labels on
dwell over the buttons.

Also considered was a more magical variant where moving onto the
vertical icon toolbar causes the sidebar to overlay into the content to
its right with full text labels.  But Semantic UI can't do this out of
box and it seems likely to be an infuriating unstable-feeling UX.