Remove history API use from legacy search page.
The fancy UI's use of the history API was inadvertently triggering the
search page logic that would trigger a reload when using the back button
on search pages.  (It wouldn't trigger on normal pages because no state
would ever be pushed so no popstate event would ever be generated.)

This will slightly regress the hacky legacy search page support, but it's
intended to be a stopgap until it's removed anyways.