Improve history use for session tabbed UI switching between files.
This also removes use of dxr.js from 'fancy' (not the legacy search) pages
which means I can probably revert the change that eliminated dxr.js' use of
the history API.

It does appear this broke line-number clicking, but that really wants to be
converted to my proposed semantic anchoring overkill.  Or at least that's a
good way to handwave away the breakage.