Fix pretty corruption by eliminating (inconsistent) pretty demangling
router.py was previously demangling possibly C++ symbols on the fly when
it encountered raw symbols.  sorch could end up returning the raw value as
a pretty value incorrectly because of a failure to demangle in a specific
case.  Rather than making that demangle as well, I've removed the demangling
and emission of pretty in this way since the information is now available
directly from the "meta" record for a given symbol as sourced from the
analysis files, rather that depending on round-tripping of demangling.